Move localizable enums/date into locale_strings
